+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.15.0
  * Fix compactCurrency to correctly use passed-in symbol.
  * A tweak to the way we retry on DateTime.asDate to compensate for a VM bug.
+ * Update CLDR version to 30.
 
 ## 0.14.0
  * MAJOR BREAKING CHANGE! Remove message extraction and code generation into a